% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.fsPDA.R
\name{est.fsPDA}
\alias{est.fsPDA}
\title{Forward-selected Panel Data Approach}
\usage{
est.fsPDA(treated, control, treatment_start, date = NULL, lrvar_lag = NULL)
}
\arguments{
\item{treated}{Numeric. A T-by-1 vector of time series of the treated units.}

\item{control}{Numeric. A T-by-N matrix. Each column is the times series of a control unit.}

\item{treatment_start}{An integer specifying the period treatment starts.
We suggest setting it between T/2 and T to ensure enough pre-treatment observations.}

\item{date}{Date or numeric. A T-by-1 vector of date class or any meaningful numerical sequence.
The default option \code{NULL} means \code{1:length(treated)} is used.}

\item{lrvar_lag}{A non-negative integer specifying the maximum lag with Bartlett kernel
for the Newey-West long-run variance estimator.
The default option \code{NULL} means \code{floor((length(treated)-treatment_start+1)^(1/4))} is used.
It cannot be larger than \code{floor(sqrt(length(treated)-treatment_start+1))}.}
}
\value{
A list is returned of the class "fsPDA". It contains the following components:
\item{select}{A list containing the selected control units,
  where \code{dim} is the number of selected units,
  \code{control} is the vector indicates the selected units,
  \code{coef} contains the coefficient estimates,
  and \code{RSquared} is the in-sample R-Squared.}
\item{in_sample}{A data frame with in-sample fitted values.}
\item{out_of_sample}{A data frame with out-of-sample counterfactual predicts and treatment effect estimates.}
\item{ATE}{A numeric vector containing estimate of average treatment effect (ATE),
  its long-run variance, t-statistic, and p-value to test if ATE is statistically 0.}
}
\description{
Estimation and inference by the forward-selected panel data approach in Shi and Huang (2021).
}
\examples{
library(fsPDA)

# Example of China Import
data("china_import")
date_import <- names(china_import$treated)
result <- est.fsPDA(treated = china_import$treated, control = china_import$control,
                    treatment_start = which(date_import == china_import$intervention_time),
                    date = as.Date(paste(substr(date_import,1,4), "-", substr(date_import, 5, 6), "-01", sep="")))
print(result)
plot(result, tlab = "Year", ylab = "Monthly Growth Rate")

# Example of HCW
data("HCW")
result <- est.fsPDA(treated = HCW$panel[,1], control = HCW$panel[,-1],
                    treatment_start = HCW$T1+1,
                    date = as.Date(paste(substr(HCW$quarter, 1, 4), "-", (as.numeric(substr(HCW$quarter, 6, 6))-1)*3+1, "-1", sep="")))
print(result)
plot(result, tlab = "Year", ylab = "Real GDP Growth Rate")

}
\references{
Zhentao Shi and Jingyi Huang (2021): "Forward-Selected Panel Data Approach for Program Evaluation,"
  forthcoming at the Journal of Econometrics, arXiv: 1908.05894
}
