% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsPDA.R
\docType{data}
\name{china_import}
\alias{china_import}
\title{Chinese Import Data}
\format{
\itemize{
 \item\strong{Frequency:} {Monthly}
 \item\strong{Date Range:} {2010:Feb -- 2015:Dec}
 \item\code{treated} {Time series of luxury watch import, under the official category
 name "watches with case of, or clad with, precious metal". Names are in the format "yyyymm".}
 \item\code{control} {A matrix of commodities as the control units. The rowname is "yyyymm", and the column name
 is the identity of the commodity.}
 \item\code{intervention_time} {Character. The month when the treatment intervention started, formated in "yyyymm".}
}
}
\source{
United Nations Comtrade Database (\url{https://comtrade.un.org/}).
This database provides detailed statistics for international commodity trade.

Pre-treatment: 2010:Feb -- 2012:Dec (35 months). Post-treatment: 2013:Jan -- 2015:Dec (36 observations).

China's monthly import data are unavailable after September of 2012.
To make our data consistent,
we sum the other countries' export values to China, and take them as import values before and after September of 2012.
}
\usage{
data(china_import)
}
\description{
This dataset is described in Section 5 of Shi and Huang (2021).
}
\references{
Zhentao Shi and Jingyi Huang (2021): "Forward-Selected Panel Data Approach for Program Evaluation,"
  forthcoming at the Journal of Econometrics, arXiv: 1908.05894
}
\keyword{datasets}
