% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/china_import.R
\docType{data}
\name{china_import}
\alias{china_import}
\title{Chinese import data}
\format{
\itemize{
 \item\strong{Release:} {United Nations}
 \item\strong{Frequency:} {Annual}
 \item\strong{Date Range:} {20xx -- 20xx}
}
}
\source{
United Nations Comtrade Database \url{https://comtrade.un.org/}
}
\usage{
data(china_import)
}
\description{
Chinese import data
}
\section{Description}{

This dataset is described in Section 5 of Shi and Huang (2021).
}

\references{
Zhentao Shi and Jingyi Huang (2021): "Forward-Selected Panel Data Approach for Program Evaluation,"
  forthcoming at the Journal of Econometrics, arXiv: 1908.05894
}
\keyword{datasets}
