% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsPDA.R
\docType{package}
\name{fsPDA}
\alias{fsPDA}
\title{fsPDA: Package for the Forward-selected Panel Data Approach}
\description{
Estimation and inference by the forward-selected panel data approach in Shi and Huang (2021).
}
\details{
\code{est.fsPDA} is the main function. It generates a \code{fsPDA} object which can work with
the generic methods \code{plot}.
}
\references{
Zhentao Shi and Jingyi Huang (2021): "Forward-Selected Panel Data Approach for Program Evaluation,"
  forthcoming at the Journal of Econometrics, arXiv: 1908.05894
}
