% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsPDA.R
\name{fsPDA}
\alias{fsPDA}
\title{Forward-selected panel data approach}
\usage{
fsPDA(treated, control, treatment_start, date = NULL, lrvar_lag = NULL)
}
\arguments{
\item{treated}{Numeric. A T-by-1 vector of time series of the treated units.}

\item{control}{Numeric. A T-by-N matrix. Each column is the times series of a control unit.}

\item{treatment_start}{An integer specifying the period treatment starts.
We suggest setting it between $T/2$ and $T$ to ensure enough pre-treatment observations.}

\item{date}{Date or numeric. A $T$-dimensional vector of date class or any meaningful numerical sequence.
The default option `NULL` means `1:length(treated)` is used.}

\item{lrvar_lag}{A non-negative integer specifying the maximum lag with Bartlett kernel
for the Newey-West long-run variance estimator.
The default option `NULL` means `floor((length(treated)-treatment_start+1)^(1/4))` is used.
It cannot be larger than `floor(sqrt(length(treated)-treatment_start+1))`.}
}
\value{
A list is returned of the class `fsPDA`. It contains the following components:
\item{select}{A list containing the selected control units,
  where `dim` is the number of selected units,
  `control` is the vector indicates the selected units,
  `coef` contains the coefficient estimates,
  and `RSquared` is the in-sample $R^2$.}
\item{in_sample}{A data frame with in-sample fitted values.}
\item{out_of_sample}{A data frame with out-of-sample counterfactual predicts and treatment effect estimates.}
\item{ATE}{A numeric vector containing estimate of average treatment effect (ATE),
  its long-run variance, t-statistic, and p-value to test if ATE is statistically 0.}
\item{plot}{A `ggplot` object. Some post-plot arguments of `ggplot` can be added additionally, for example, `labs`.}
}
\description{
Function implementing forward-selected panel data approach in Shi and Huang (2021)
}
\examples{
library(fsPDA)
data("testData")

treated = testData$treated
control = testData$control
intervention_time = testData$intervention_time

result=fsPDA(treated, control,
            treatment_start = which(names(treated) == intervention_time),
            date = as.Date(paste(substr(names(treated),1,4), "-", substr(names(treated), 5, 6), "-01", sep="")))

print(result)

}
\references{
Zhentao Shi and Jingyi Huang (2001): "Forward-Selected Panel Data Approach for Program Evaluation,"
  Journal of Econometrics, arXiv: 1908.05894
}
