% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsPDA.R
\docType{data}
\name{HCW}
\alias{HCW}
\title{HCW Data}
\format{
\itemize{
  \item\code{panel} {A matrix of real GDP growth rate of many countries.}
  \item\code{quarter} {Character. A time sequence related to \code{panel} labeled by year and quarter.}
  \item\code{T1} {An Integer indicates the length of pre-treated period.}
}
}
\usage{
data(HCW)
}
\description{
This is the real GDP growth rate data used in Hsiao, Ching, and Wan (2012).
}
\references{
Hsiao, C., S. H. Ching, and S. K. Wan (2012):
  “A panel data approach for program evaluation: measuring the benefits of political and economic integration of Hong Kong with mainland China,”
  Journal of Applied Econometrics, 27(5), 705–740.
}
\keyword{datasets}
