% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fsPDA.R
\name{plot.fsPDA}
\alias{plot.fsPDA}
\title{Plot Counterfactual Prediction}
\usage{
\method{plot}{fsPDA}(
  x,
  tlab = NULL,
  ylab = NULL,
  title = NULL,
  point = TRUE,
  legend.position = "bottom"
)
}
\arguments{
\item{x}{An object of class the "fsPDA".}

\item{tlab}{The label of the time (horizontal) axis.}

\item{ylab}{The label of the value (vertical) axis.}

\item{title}{The text for the title.}

\item{point}{Logical. Should a layer of points be included in the plot?}

\item{legend.position}{The position of legends ("none", "left", "right", "bottom", "top").}
}
\description{
Plot method for objects of class "fsPDA".
}
\note{
"ggplot2" package needs to be installed as dependency.
}
\seealso{
\link{est.fsPDA} for examples.
}
